[FAB-8658] peer mock functions to separate file
Production and test should not be co-mingled in the same files.

The mock initialization functions are currently referenced by code
outside of the peer package and can't be moved to an _test.go file
right now without introducing compilation errors.
